% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allomCarbon.R
\docType{data}
\name{allomCarbon}
\alias{allomCarbon}
\title{Information about output$allomCarbon dataframe}
\format{
A data frame with 7 variables:
\describe{
  \item{X}{The growth ring number of the disc at ground}
  \item{treeage}{The age class of a tree growths, year}
  \item{abovegroundB}{The aboveground biomass of a sampled tree, kg}
  \item{belowgroundB}{The belowground biomass of a sampled tree, kg}
  \item{totalB}{The total tree biomass of a sampled tree, kg}
  \item{abovegroundC}{The aboveground carbon storage of a sampled tree, kg}
  \item{belowgroundC}{The belowground carbon storage of a sampled tree, kg}
  \item{totalC}{The total tree carbon storage of a sampled tree, kg}
}
}
\usage{
allomCarbon
}
\description{
A dataframe containing the list of output variables and their description.
}
\examples{
head(allomCarbon)
}
\keyword{datasets}
