% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BEFdata}
\alias{BEFdata}
\title{BEFdata}
\format{
An object of class \code{data.frame} with 1 rows and 4 columns.
}
\usage{
BEFdata
}
\description{
Just test number(Description)
}
\examples{
head(BEFdata)
}
\keyword{datasets}
