% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_parameters.R
\docType{data}
\name{d_parameters}
\alias{d_parameters}
\title{Information about input variables in parameters of allometric models}
\format{
A data frame with 6 rows and 4 variables:
\describe{
  \item{tissues}{The tree tissues including stem, branch, leaf, root, and total tree}
  \item{a}{The parameter a in the allometric model ln(Bi)=a+b×ln(DBH2H)}
  \item{b}{The parameter b in the allometric model ln(Bi)=a+b×ln(DBH2H)}
  \item{Cconcentration}{The tree carbon concentration, kg C kg-1}
}
}
\usage{
d_parameters
}
\description{
A dataset containing the list of input variables and their description.
    Note: If a user uses the StemAnalysis package to estimate a sampled tree
    carbon by allometric models, a parameters dataset should be provided.
    Owing the allometric model varies among tree species, the users should
    input the parameter data of the allometric models for the correspondingly
    tree species.
}
\examples{
head(d_parameters)
}
\keyword{datasets}
