% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataframe1}
\alias{dataframe1}
\title{dataframe1}
\format{
An object of class \code{data.frame} with 5 rows and 4 columns.
}
\usage{
dataframe1
}
\description{
Just test dataframe(Description)
}
\examples{
head(dataframe1)
}
\keyword{datasets}
