% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataframe2}
\alias{dataframe2}
\title{dataframe2}
\format{
An object of class \code{data.frame} with 1 rows and 4 columns.
}
\usage{
dataframe2
}
\description{
Just test dataframe(Description)
}
\examples{
head(dataframe2)
}
\keyword{datasets}
