% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameterdata}
\alias{parameterdata}
\title{parameterdata}
\format{
An object of class \code{data.frame} with 5 rows and 4 columns.
}
\usage{
parameterdata
}
\description{
Just test number(Description)
}
\examples{
head(parameterdata)
}
\keyword{datasets}
