% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_allomCarbon.R
\docType{data}
\name{r_allomCarbon}
\alias{r_allomCarbon}
\title{Information about output$allomCarbon dataframe}
\format{
A data frame with 11 variables:
\describe{
  \item{treeage}{The age class of a tree growths, year}
  \item{stemB}{The stem biomass of a sampled tree, kg}
  \item{branchB}{The branch biomass of a sampled tree, kg}
  \item{leafB}{The leaf biomass of a sampled tree, kg}
  \item{rootB}{The root biomass of a sampled tree, kg}
  \item{totalB}{The total tree biomass of a sampled tree, kg}
  \item{stemC}{The stem carbon storage of a sampled tree, kg}
  \item{branchC}{The branch carbon storage of a sampled tree, kg}
  \item{leafC}{The leaf carbon storage of a sampled tree, kg}
  \item{rootC}{The root carbon storage of a sampled tree, kg}
  \item{totalC}{The total tree carbon storage of a sampled tree, kg}
}
}
\usage{
r_allomCarbon
}
\description{
A dataframe containing the list of output variables and their description.
}
\examples{
head(r_allomCarbon)
}
\keyword{datasets}
