% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_StemGrowth.R
\docType{data}
\name{r_StemGrowth}
\alias{r_StemGrowth}
\title{Information about output$StemGrowth dataframe}
\format{
A data frame with 10 variables:
\describe{
  \item{stemdj}{The age class of a tree growths, year}
  \item{DBHt}{The tree diameter at breast height, cm}
  \item{Height}{The tree height, m}
  \item{Volume}{The tree stem volume, m3}
  \item{AnincreD}{The mean annual increment of diameter at breast height, cm}
  \item{AvincreD}{The current annual increment of diameter at breast height, cm}
  \item{AnincreH}{The mean annual increment of tree height, m}
  \item{AvincreH}{The current annual increment of tree height, m}
  \item{AnincreV}{The mean annual increment of tree stem volume, m3}
  \item{AvincreV}{The current annual increment of tree stem volume, m3}
}
}
\usage{
r_StemGrowth
}
\description{
A dataframe containing the list of output variables and their description.
}
\examples{
head(r_StemGrowth)
}
\keyword{datasets}
