% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_volumeCarbon.R
\docType{data}
\name{r_volumeCarbon}
\alias{r_volumeCarbon}
\title{Information about output$volumeCarbon dataframe}
\format{
A data frame with 9 variables:
\describe{
  \item{treeage}{The age class of a tree growths, year}
  \item{stemB}{The stem biomass of a sampled tree, kg}
  \item{abovegroundB}{The aboveground biomass of a sampled tree, kg}
  \item{belowgroundB}{The belowground biomass of a sampled tree, kg}
  \item{totalB}{The total tree biomass of a sampled tree, kg}
  \item{stemC}{The stem carbon storage of a sampled tree, kg}
  \item{abovegroundC}{The aboveground carbon storage of a sampled tree, kg}
  \item{belowgroundC}{The belowground carbon storage of a sampled tree, kg}
  \item{totalC}{The total tree carbon storage of a sampled tree, kg}
}
}
\usage{
r_volumeCarbon
}
\description{
A dataframe containing the list of output variables and their description.
}
\examples{
head(r_volumeCarbon)
}
\keyword{datasets}
