% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemanalysis.R
\name{stemanalysism}
\alias{stemanalysism}
\title{Reconstructing tree growth and carbon accumulation with stem analysis data}
\usage{
stemanalysism(
  xtree,
  stemgrowth = FALSE,
  treecarbon = FALSE,
  HDmodel = FALSE,
  stemdata,
  parameterdata,
  BEFdata
)
}
\arguments{
\item{xtree}{Xtree is the tree number (Treeno), which isused to choose the
target tree to be analyzed}

\item{stemgrowth}{If stemgrowth is 'TRUE', stem growth profile and growth
trends in terms of diameter at breast height (DBH), tree height, and
stem volume will be showed in a graph}

\item{treecarbon}{If treecarbon is 'TRUE', total tree biomass and carbon
storage will be estimated by allometric models and volume model. In
addition, although treecarbon is 'TRUE', the estimation of tree biomass
and carbon storage by allometric models will skip if data 'parameterdata'
is missing, and the same is true for the estimation by volume model if
data 'BEFdata' is missing}

\item{HDmodel}{If HDmodel is 'TRUE', height-diameter relationship will be
fitted with nonlinear models and showed the fitted results in a graph}

\item{stemdata}{Stemdata is the stem analysis data that has been inputted}

\item{parameterdata}{Parameterdata is the parameter data of allometric
models that can be optionally inputted}

\item{BEFdata}{BEFdata is the biomass estimation factor data of volume model
that can be optionally inputted by users}
}
\description{
Reconstructing tree growth and carbon accumulation with stem analysis data
}
\examples{

library(StemAnalysis)

# Load the data sets
data(stemdata)
data(parameterdata)

# To calculating tree growth and carbon accumulation with input data sets
stemanalysism(xtree = 8, stemdata = stemdata)

# If the graph of stem growth profile and growth trends is needed
stemanalysism(xtree = 8, stemgrowth = TRUE, treecarbon = TRUE,
    stemdata = stemdata, parameterdata = parameterdata)

}
