% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumePardata.R
\docType{data}
\name{volumePardata}
\alias{volumePardata}
\title{Information about input variables in biomass expansion factor for volume model}
\format{
A data frame with 5 rows and 7 variables:
\describe{
  \item{X}{The line number}
  \item{DBH}{The tree diameter of breast height, cm}
  \item{factors}{The tree biomass estimation factors including BCF and RSR}
  \item{a}{The parameter a in the estimation model of BCF (ln(BCF)=ln(a)+bﾃ様n(DBH)+cﾃ様n(H)) and RSR (ln(RSR)=ln(a)+bﾃ様n(DBH)+cﾃ様n(H))}
  \item{b}{The parameter b in the estimation model of BCF and RSR}
  \item{c}{The parameter c in the estimation model of BCF and RSR}
  \item{Cconcentration}{The total tree carbon concentration, kg C kg-1}
}
}
\usage{
volumePardata
}
\description{
A dataset containing the list of input variables and their description.
    Note: If a user uses the StemAnalysis package to estimate a sampled tree
    carbon by volume model, a parameters dataset for the factors (BCF and RSR)
    should be provided. Owing the parameters of BCF and RSR vary among tree
    species, the users should input the parameters data for the
    correspondingly tree species
}
\examples{
head(volumePardata)
}
\keyword{datasets}
