% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stemdata}
\alias{stemdata}
\title{stemdata}
\format{
An object of class \code{data.frame} with 97 rows and 18 columns.
}
\usage{
stemdata
}
\description{
Just test number(Description)
}
\examples{
head(stemdata)
}
\keyword{datasets}
